Rust: Refactor RoadBuilder and LaneBuilder
These are two quite big structs, so we begin a refactor by moving them out of mod to allow for better increments
